Moving trimming service registrations to root Startup
